+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/OAuth2Client.kt
@@ -79,7 +79,7 @@ class OAuth2Client(
      * of [IssuerMetadata.credentialIssuer]
      * @param issuerState for OID4VCI flows the value from [CredentialOfferGrantsAuthCode.issuerState]
      * @param audience for PAR the value of the `issuer` of the Authorization Server
-     * @param wrapAsPar whether to wrap the request as a PAR (i.e. a signed JWS)
+     * @param wrapAsJar whether to wrap the request as a JAR (i.e. a signed JWS with the authn request as payload)
      */
     suspend fun createAuthRequest(
         state: String,
@@ -88,7 +88,7 @@ class OAuth2Client(
         resource: String? = null,
         issuerState: String? = null,
         audience: String? = null,
-        wrapAsPar: Boolean = true,
+        wrapAsJar: Boolean = true,
     ) = AuthenticationRequestParameters(
         responseType = GRANT_TYPE_CODE,
         state = state,
@@ -100,12 +100,12 @@ class OAuth2Client(
         redirectUrl = redirectUrl,
         codeChallenge = generateCodeVerifier(state),
         codeChallengeMethod = CODE_CHALLENGE_METHOD_SHA256
-    ).wrapIfNecessary(wrapAsPar, audience)
+    ).wrapIfNecessary(wrapAsJar, audience)
 
-    private suspend fun AuthenticationRequestParameters.wrapIfNecessary(wrapAsPar: Boolean, audience: String?) =
-        if (signPushedAuthorizationRequest != null && wrapAsPar) wrapInPar(signPushedAuthorizationRequest, audience) else this
+    private suspend fun AuthenticationRequestParameters.wrapIfNecessary(wrapAsJar: Boolean, audience: String?) =
+        if (signPushedAuthorizationRequest != null && wrapAsJar) wrapInJar(signPushedAuthorizationRequest, audience) else this
 
-    private suspend fun AuthenticationRequestParameters.wrapInPar(
+    private suspend fun AuthenticationRequestParameters.wrapInJar(
         signPushedAuthorizationRequest: SignJwtFun<AuthenticationRequestParameters>,
         audience: String?,
     ) = AuthenticationRequestParameters(